+++ tests/python/csp.py
@@ -11,9 +11,9 @@ def main():
     # demo for calculating the CSP filters
 
     n_ep, n_ch, n_times = (2, 2, 4)
-    labels = np.array([0, 1])
+    labels = np.array([0, 1]).astype('float64') 
     
-    data = np.array([[[6, 3, 1, 5], [3, 0, 5, 1]], [[1, 5, 6, 2], [5, 1, 2, 2]]])  
+    data = np.array([[[6, 3, 1, 5], [3, 0, 5, 1]], [[1, 5, 6, 2], [5, 1, 2, 2]]]).astype('float64')
     
     filters, ev = DataFilter.get_csp(data, labels);
     # filters =
